backend/requirements.txt
# Backend dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
sqlalchemy==2.0.23
alembic==1.13.1
google-generativeai==0.3.2
pydantic[email]==2.5.0
email-validator==2.1.0
aiofiles==23.2.1
python-dotenv==1.0.0

# Flask alternative (optional)
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-jwt-extended==4.6.0

# Development
pytest==7.4.3
httpx==0.25.2
